move config into separate folder/ file
